+++ lib/td_dq/search/indexer.ex
@@ -5,7 +5,6 @@ defmodule TdDq.Search.Indexer do
 
   alias Elasticsearch.Index
   alias Elasticsearch.Index.Bulk
-  alias Jason, as: JSON
   alias TdCache.Redix
   alias TdDq.Rules.Implementations.Implementation
   alias TdDq.Rules.Rule
@@ -39,7 +38,7 @@ defmodule TdDq.Search.Indexer do
     {:ok, _} =
       mappings
       |> Map.put(:index_patterns, "#{index}-*")
-      |> JSON.encode!()
+      |> Jason.encode!()
       |> put_template(index)
 
     Index.hot_swap(Cluster, index)